ew_agents/requirements.txt
# ElectionWatch Agents - Google ADK Requirements
# Core ADK dependencies
google-adk==1.5.0
google-cloud-aiplatform[agent_engines]>=1.95.1

# Agent tools dependencies
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
pymongo>=4.0.0
neo4j==5.22.0

# ML and NLP dependencies
llama-index==0.10.53
llama-index-embeddings-huggingface
transformers>=4.30.0
torch>=2.0.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Optional enhanced features
langchain-community==0.2.10
langgraph==0.1.13

# HTTP and async
aiohttp>=3.8.0
requests>=2.28.0 